Use a trie to look up URL Maps
This implementation isn't much space-efficient, but is sufficiently fast and
scales better than the previous loop-through-all-the-maps-doing-strncmp
approach.

Lookup is O(k) instead of approx. O(ka * n) in the previous approach; k
being the key length, ka being the average key length and n being the number
of entries in the URL map.  Since prefix matching is often used, k is also
often very small, which is also a win.