Better keep-alive implementation
This uses a wait-free queue that actually manages each client socket's
time-to-live, instead of the previous hey-it's-timeout-kill-everyone
approach that was being used previously.  Also, there is better request
socket management, so a socket won't be closed twice (which, due to a
possible race condition, a thread could kill other thread's socket in
mid-air).

Apachebench still does not like this, though. But the strace log looks
promising, will have to take a look later.

Also, since the keep-alive implementation is working, file serving is now
broken!  Fear not, because it is now time to revamp that file serving cruft
so that it doesn't block on sendfile().  But later, I'm late for work.