Make sure file is closed even if coroutine is aborted
The lwan_sendfile() function can yield from the coroutine -- and it might be
killed, for instance if the connection is closed.  Using the deferred table
from the coroutine, we make sure the file is closed if the coroutine is
closed at the right moment regardless if the socket has been closed or not.