Get rid of _create_temporary_cache_entry()
If the cache would block on read, just yield from the coroutine to try
again. Getting rid of one less common way of serving a file means less
tested code is not around to be a problem later on.

Next step in cleaning file serving is to cache file not found errors, since
cache misses are expensive.