Register modules on demand
Instead of registering all modules manually in `lwan_init()`, they're
looked up on demand by `lwan_module_find()`, by using
`find_handler_symbol()`.

Due to Lwan using LTO when linking and unused symbols defined in
library archive files not being considered for inclusion in the final
executable, there has been a change in how the executables are linked:
it's now required to pass the `--whole-archive` option to the linker
(and disable it after specifying the `lwan-common` library, of course).
This works around the behavior and includes the visible symbols in the
final executable.

At the moment this requires only a module to export a function prefixed
with `lwan_module_`, followed by the module name.  It should return a
`const lwan_module_t *` structure.  Eventually this should move so that
these functions are in a separate section in the ELF file, to avoid
possible name clashes (only functions located in that section will be
considered by Lwan when looking up modules).

The lookup is cached in the same hash table used previously.

The major benefit of this change is that creating a module is pretty
straightforward; there's no need to change Lwan itself so it can be
registered.  Also, if a module is only available if a certain external
library is available (such as Lua), nothing else has to be done: lookup
will still fail in runtime.