Plug memory leak after moving coro_defer to an array
Previously, coro_malloc() would allocate a chunk of memory sufficiently
large to hold a struct coro_defer and the requested amount of memory;
the callback to free the struct was an empty function, since the linked
list node would be freed with free() anyway.  With the move to an array
rather than a linked list, the deferred callback should be a call to
free().