Fix double free when iterating through isolated config
find_section_end() modifies the buffer, and strbuf_init() was still
being called with the old buffer pointer -- and which would be freed
by the isolated section (when closed) as well as the non-isolated
section.

Call strbuf_init() on the isolated strbuf as soon as the copy from
the original configu struct is performed. This ensures that
find_section_end() operates on a pristine buffer.